changed int to float and some settings in physics engine, glitching occures a lot less now :D

+++ Dimitri/Engines/Graphics/src/Models/Center.h
@@ -30,12 +30,12 @@ namespace Graphics {
 			/**
 			* \brief The x coordinate
 			*/
-			int x;
+			float x;
 
 			/**
 			* \brief The y coordinate
 			*/
-			int y;
+			float y;
 		};
 	}
 }
